Sema: Split off RequiremehtMatch.h from TypeCheckProtocol.h
